+++ pyanimelist/__init__.py
@@ -1,8 +1,8 @@
-__version__ = "1.2.2"
-__author__ = 'Byron Vanstien'
 __license__ = 'MIT'
-__copyright__ = 'Copyright 2016-2017 Byron Vanstien'
+__version__ = "1.2.2"
 __title__ = 'pyanimelist'
+__author__ = 'Byron Vanstien'
+__copyright__ = 'Copyright 2016-2017 Byron Vanstien'
 
 from .errors import *
 from .client import *