+++ pyanimelist/abstractions.py
@@ -1,10 +1,14 @@
+from typing import List
+from datetime import datetime
+
+
 class Titles(object):
 
     """
     Abstraction class for the three versions of titles myanimelist keeps.
     """
 
-    def __init__(self, jp: str, english: str, synonyms: list = None):
+    def __init__(self, jp: str, english: str, synonyms: List[str] = None):
         self.jp = jp
         self.english = english
         self.synonyms = synonyms or []
@@ -16,6 +20,6 @@ class Dates(object):
     Abstraction class for the dates myanimelist stores
     """
 
-    def __init__(self, start: str, end: str):
+    def __init__(self, start: datetime, end: datetime):
         self.start = start
         self.end = end