+++ pyanimelist/scraper.py
@@ -1,0 +1,9 @@
+from util import fetch_url
+
+
+class Scraper(object):
+    """
+    A Scraper class which facilitates all the requests and scraping
+    """
+    def __init__(self):
+        pass