+++ docs/source/whats_new.rst
@@ -1,6 +1,11 @@
+.. currentmodule:: pyanimelist
+
 What's new!
 -----------
 
-.. note::
+This page keeps a human readable list of what has changed throughout the versions, starting from v1.2.3
 
-   This will be added from onwards with each library version bump.
+v1.2.3
+------
+* pyanimelist.errors.InvalidCredentials now inherits from ResponseError
+* Switch from strings to datetime objects on :class:`pyanimelist.abstractions.Dates`